Use DateTime::createFromFormat
