+++ tests/unit/Trace/SpanTest.php
@@ -181,9 +181,9 @@ class SpanTest extends TestCase
     {
         return [
             ['startTime', 1490737410, '2017-03-28T21:43:30.000000000Z'],
-            ['startTime', 1490737450.4843, '2017-03-28T21:44:10.484299000Z'],
+            ['startTime', 1490737450.484299, '2017-03-28T21:44:10.484299000Z'],
             ['endTime', 1490737410, '2017-03-28T21:43:30.000000000Z'],
-            ['endTime', 1490737450.4843, '2017-03-28T21:44:10.484299000Z'],
+            ['endTime', 1490737450.484299, '2017-03-28T21:44:10.484299000Z'],
         ];
     }
 }
